+++ app/frontend/src/views/minesattestations/InspectionDashboard.vue
@@ -22,7 +22,7 @@ export default {
         : undefined;
     },
     resource() {
-      return AppClients.MINESATTESTATIONS;
+      return AppClients.MINESOPERATORSCREENING;
     }
   }
 };
