+++ app/src/forms/minesoperatorscreening/assets/status-assignment-email.html
